Reviewers: name the response column, and eight defects they found
FOUR REVIEWERS. A cold reader, an ML engineer, a numerical analyst and a C
systems engineer. This is what has been acted on; the rest is reported to the
author, not silently dropped.

-y NAME / --response NAME. The cold reader wrote the CSV a pandas user
exports, site,dose,age,response, and got dose regressed on age and response:
plausible coefficients, exit 0, nothing wrong on the face of it. The layout is
positional and there was no way to say otherwise. Now there is; column 1 stays
the group, the named column becomes the value, every other column is a term.
The positional default is unchanged, and the reading: line ends with the
remedy when the column was not named, so the mistake shows on the first run
rather than in the numbers.

A QUOTE ANYWHERE, not only at byte zero. `A"` and `A` became two groups,
silently, which is the outcome the guard was written to prevent and it looked
at one byte. First or last now, so O'Brien is still a name.

THE SCORING PATH threw away the diagnostic the parser had already composed and
printed the generic sentence the rewrite existed to remove. Both training
paths did it right; this one did not.

ROW COUNTERS WERE long, which is 32 bits on the Windows target now being
shipped, so every one of them overflows at 2.1 billion rows while the README
tabulates a hundred trillion. All of them, and every format that prints one,
are long long.

THE COLUMN NORM in qr.c squared the raw value, so a column past about 1.3e154
sent the accumulator to +inf, made the scaled diagonal zero, and deleted a
perfectly identified term as COLLINEAR. It now accumulates relative to a
running maximum. The rotation two screens below already used hypot() to avoid
exactly this. Storage grows by one vector per column, so the documented factor
is p^2+7p+6, not p^2+6p+5.

THE PROGRESS LINE counted rows kept, not rows read, so -g on a rare group in a
billion-row file printed nothing for the whole run. It also used a wall clock,
which can step backwards. Monotonic where there is one, and a backward step
restarts the interval instead of ending the reporting.

THE PRODUCTION LINK swept every first-level directory, so c/tests.c is in the
shipped binary today and anything dropped into bench/ or scripts/ would join
it. Scoped to c/.

release.yml refused to publish when the tag and LINEARR_VERSION disagree;
--footprint took atoi, so "abc" meant 0 and fell through to a message about a
missing coefficient table; the ill-conditioning warning advised --qr to people
already running --qr; a group name was memcpy'd out of an uninitialised
automatic.

AND A GATE THAT DIED SILENTLY. tests/cli.sh appends a probe to utils.c to
prove a real warning surfaces. When the sources moved to c/ the probe landed
in a file nothing compiles, grep matched nothing, exited 1, and `set -e` ended
the whole script with no output and no FAIL line. Both fixed: the path, and
the `|| true` that makes a broken probe fail its own assertion instead of the
run.

NOT FIXED, AND NOW WRITTEN DOWN. The residual SD from the default solver is
Cyy - b'Cxy, and on 200 rows of an exact quadratic fitted with a line it
reports 0.3718, then 0, then 11.57, then 1256, as x moves from 1e4 to 1e7. The
truth is 0.3745 throughout and the residual file written by the same run says
so. A guard was tried and withdrawn: every threshold that caught 1e6 and 1e7
also suppressed the figure on every exactly-fitting example, including the
0.02282 that Wampler1 exists to show. --qr returns 0.3745, 0.3746, 0.3736 and
0.5018 on the same four. regress.c carries the measurement and says to use it.

README claims corrected: "eleven digits" held for Norris and Longley and not
for Wampler1, which is nine and ten; "agrees with lm() to 1e-11" is 1e-6 in
the gate and two files exceed 1e-11. And bench/fit-frame.py is hand-written
Python that says so in its own docstring, while four places in the README
called its numbers pandas. Real pandas is faster and lighter and is not
measured here; the row is labelled for what it is.

+++ c/diag.h
@@ -112,13 +112,13 @@
 
 struct diag {
     int     nvars;
-    long    n;
+    long long n;
     double *s;
     double  resid_sd, response_sd;
 };
 
 struct diag_result {
-    long   rows;
+    long long rows;
     int    curved_term;   /* term whose square explains the residual, or -1 */
     double curved_t;      /* its t statistic, 0 when there is none          */
     int    curved_pow;    /* 2 or 3: which power explained the residual     */
